# basicprotocols/group_chat/api/swagger/docs/group_chat_swagger_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Group Chat Service API documentation, including database queries, group management, community management and other functions",
        "title": "Group Chat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:7568",
    "basePath": "/",
    "paths": {
        "/group-chat/group-list": {
            "get": {
                "description": "Get group list with pagination support",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group Management"
                ],
                "summary": "Get group list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User MetaId",
                        "name": "metaId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Cursor, default is 1",
                        "name": "cursor",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default is 20",
                        "name": "size",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp",
                        "name": "timestamp",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully return group list",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/group-chat/user/latest-group-list": {
            "get": {
                "description": "Get user's latest chat group list, sorted by latest chat time",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group Management"
                ],
                "summary": "Get latest chat group list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User MetaId",
                        "name": "metaId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cursor, default is 1",
                        "name": "cursor",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default is 20",
                        "name": "size",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp",
                        "name": "timestamp",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully return latest chat group list",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/group-chat/group-info": {
            "get": {
                "description": "Get detailed information of specified group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group Management"
                ],
                "summary": "Get group information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully return group information",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Group does not exist",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/group-chat/group-chat-list": {
            "get": {
                "description": "Get group chat records with timestamp pagination support",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group Management"
                ],
                "summary": "Get group chat records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User MetaId",
                        "name": "metaId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Cursor, default is 0",
                        "name": "cursor",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default is 20",
                        "name": "size",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp for pagination",
                        "name": "timestamp",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully return group chat records",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/group-chat/group-member-list": {
            "get": {
                "description": "Get group member list with pagination support",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group Management"
                ],
                "summary": "Get group member list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cursor, default is 1",
                        "name": "cursor",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default is 20",
                        "name": "size",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp",
                        "name": "timestamp",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully return group member list",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/group-chat/group-person": {
            "get": {
                "description": "Get TalkGroupPersonCollection information based on metaId and groupId to determine if user is in specified group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group Management"
                ],
                "summary": "Get group member information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User MetaId",
                        "name": "metaId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully return group member information",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
}